refactor(eur-update-service): lib.rs into separate files
