+++ crates/backend/eur-update-service/src/utils.rs
@@ -1,0 +1,135 @@
+//! Utility functions for parsing and validation
+
+use anyhow::Result;
+use semver::Version;
+
+use crate::error::UpdateServiceError;
+
+/// Parse target_arch into target and arch components
+/// e.g., "linux-x86_64" -> ("linux", "x86_64")
+pub fn parse_target_arch(target_arch: &str) -> Result<(String, String)> {
+    let parts: Vec<&str> = target_arch.split('-').collect();
+    if parts.len() < 2 {
+        return Err(anyhow::Error::from(UpdateServiceError::InvalidTargetArch(
+            target_arch.to_string(),
+        )));
+    }
+
+    let target = parts[0].to_string();
+    let arch = parts[1..].join("-"); // Handle cases like "aarch64" or multi-part arch
+
+    Ok((target, arch))
+}
+
+/// Extract version from S3 object key for the new structure
+/// Expected format: releases/{channel}/{version}/{target}/{arch}/filename
+pub fn extract_version_from_key(
+    key: &str,
+    prefix: &str,
+    target: &str,
+    arch: &str,
+) -> Option<String> {
+    if let Some(remainder) = key.strip_prefix(prefix) {
+        // Split by '/' to get path components
+        let parts: Vec<&str> = remainder.split('/').collect();
+
+        if parts.len() >= 3 {
+            let version_str = parts[0];
+            let key_target = parts[1];
+            let key_arch = parts[2];
+
+            // Check if this key is for our target platform
+            if key_target == target && key_arch == arch {
+                // Validate that this looks like a version
+                if Version::parse(version_str).is_ok() {
+                    return Some(version_str.to_string());
+                }
+            }
+        }
+    }
+    None
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_parse_target_arch() {
+        assert_eq!(
+            parse_target_arch("linux-x86_64").unwrap(),
+            ("linux".to_string(), "x86_64".to_string())
+        );
+
+        assert_eq!(
+            parse_target_arch("darwin-aarch64").unwrap(),
+            ("darwin".to_string(), "aarch64".to_string())
+        );
+
+        assert_eq!(
+            parse_target_arch("windows-x86_64").unwrap(),
+            ("windows".to_string(), "x86_64".to_string())
+        );
+
+        assert!(parse_target_arch("invalid").is_err());
+    }
+
+    #[test]
+    fn test_extract_version_from_key() {
+        let prefix = "releases/nightly/";
+        let target = "linux";
+        let arch = "x86_64";
+
+        assert_eq!(
+            extract_version_from_key(
+                "releases/nightly/1.0.0/linux/x86_64/bundle.AppImage.tar.gz",
+                prefix,
+                target,
+                arch
+            ),
+            Some("1.0.0".to_string())
+        );
+
+        assert_eq!(
+            extract_version_from_key(
+                "releases/nightly/1.2.3-beta.1/linux/x86_64/signature",
+                prefix,
+                target,
+                arch
+            ),
+            Some("1.2.3-beta.1".to_string())
+        );
+
+        // Different target should return None
+        assert_eq!(
+            extract_version_from_key(
+                "releases/nightly/1.0.0/darwin/x86_64/bundle.app.tar.gz",
+                prefix,
+                target,
+                arch
+            ),
+            None
+        );
+
+        // Invalid version should return None
+        assert_eq!(
+            extract_version_from_key(
+                "releases/nightly/invalid-version/linux/x86_64/file",
+                prefix,
+                target,
+                arch
+            ),
+            None
+        );
+    }
+
+    #[test]
+    fn test_version_comparison() {
+        let current = Version::parse("1.0.0").unwrap();
+        let newer = Version::parse("1.0.1").unwrap();
+        let older = Version::parse("0.9.9").unwrap();
+
+        assert!(newer > current);
+        assert!(older < current);
+    }
+}
